# HomeWork_12/ClinicService.Shared/Entity/Pet.cs
namespace ClinicService.Shared.Entity
{
    public class Pet
    {
        public int PetId { get; set; }

        public int ClientId { get; set; }

        public string Name { get; set; }

        public DateTime Birthday { get; set; }
    }
}